Auto Sync
Added an auto sync feature so the program will check for changes and carry them out every 15 seconds from the point of last check completion. The direction chosen is based on the direction of the last modified change record.

This feature will not work if only the path is updated.