Unit Tests VII
Added the start of the Document Service tests.

Need to devise a list of the methods that can't be Unit Tested.

Need to count the total number of methods & method result variations so I can get a test coverage percentage.

Look at making a widget to sit on the Read Me.