GitHub Service II
Added a run history table to the ERD.

Added a new app setting for workflows.

Created the GitHub converter to convert labels to issue types.

Altered the name of the repository property on the Workflow Run Model.

Updated the data type of the Id property in the Workflow Run Model and Release Model.

Finished the GetIssues method.

Created the GetCommits method.

Created the GetPullRequests method.

Created the GetWorkflowRuns method.

Created the GetReleases method.