Database Service XI
Created the LogRun sql.

Created the LogRun method.

Hooked the LogRun method to the ApplicationService.

Fixed logging lines.

Added additonal logging for sending of command.

Changed the list to order by the GitHub id.

Added a new property to UserModel for the date of the commit.

Ordered commit list by the date of the commit.

Fixed an issue where the issue doesn't have a type.

Fixed an issue where the issue doesn't have a solved date.