Database Service XII
Added a property onto the IssueModel to filter out pull requests obtained from the issue endpoint.

Fixed logging bugs in GitHubService.

Future proofed the OutputPullRequest method by passing 01/01/1900 if the closed or merged date is null.

Future proofed the OutputPullRequest method by passing undefined if the Type is null.

Future proofed the OutputRelease method by passing 01/01/1900 if the published date is null.

Fixed the CreateAggregates method to correctly log created and solved numbers.

Altered the CreateAggregates method to order the list by date.

Altered the CreateAggregates method to fill in the missing dates between aggregate counts.

Altered the LogRun method to take a total count of the workflow runs pulled from the API instead of the number of workflows.

All tested, final test with multiple repositories.