Unit Test V
Updated the application service configuration to accept no workflows.

Updated the Unit Tests to accept the same.

Added comments to the ApplicationServiceTest class.

Setup a Unit Test Database.

Created the DatabaseServiceTest class.

Added the Setup method.

Added the tests for the GestLastRunDate method.

Added the tests for the GetIssues method.