Upload Download Issue
Changed the code to remove the merge option where applicable. If it hasn't been downloaded, then "Up Stream" is not available. If it hasn't been uploaded, then "Down Stream" isn't an option.

Changed the code of the mass update buttons to check if the combo box on the row contains the option of that mass update.