New Folders
Changed the list of folders to include the id of the parent folder in the name.

Changed the CheckFolders method to check whether a folder exists with the specified parent.

Changed the CheckFolders method  to return the folder id of the parent of the last folder in the given list.

Changed the CreateFile and MoveFile methods to find the folder based on the generated folderStoreValue.

Changed the CreateFolder method to return the id of new folder.

Renamed the Folders key value pair list to FolderStore.

Added code to reset the FolderStore when GetData is called.